+++ config/config.go
@@ -30,4 +30,5 @@ func setDefaultValues() {
 	viper.SetDefault(settings.CommentWidthKey, settings.CommentWidthDefault)
 	viper.SetDefault(settings.IndentSizeKey, settings.IndentSizeDefault)
 	viper.SetDefault(settings.PreserveRightMarginKey, settings.PreserveRightMarginDefault)
+	viper.SetDefault(settings.HighlightHeadlinesKey, settings.HighlightHeadlinesDefault)
 }