+++ constants/settings/settings.go
@@ -19,4 +19,10 @@ const (
 	PreserveRightMarginDefault     = false
 	PreserveRightMarginDescription = "Shortens replies so that the total length, including indentation, is the same as " +
 		"the comment width. "
+	HighlightHeadlinesName        = "Highlight Headlines"
+	HighlightHeadlinesKey         = "CLX_HIGHLIGHT_HEADLINES"
+	HighlightHeadlinesDefault     = 1
+	HighlightHeadlinesDescription = "Highlights headlines containing the text Show HN, Ask HN, Tell HN and " +
+		"Launch HN. Can be set to 0 (No highlighting), 1 (inverse highlighting) and 2 (colored highlighting). Y " +
+		"Combinator Startup labels are colorized in both option 1 and 2."
 )