+++ submission/formatter/formatter.go
@@ -6,14 +6,27 @@ import (
 	"strings"
 )
 
-func GetMainText(title string, domain string) string {
-	return formatTitle(title) + formatDomain(domain)
+const (
+	black = "#0c0c0c"
+)
+
+func GetMainText(title string, domain string, mode int) string {
+	return formatTitle(title, mode) + formatDomain(domain)
 }
 
-func formatTitle(title string) string {
-	title = formatShowAndTell(title)
-	title = formatYCStartups(title)
-	return title
+func formatTitle(title string, mode int) string {
+	switch mode {
+	case 1:
+		title = labelShowAndTell(title)
+		title = labelYCStartups(title)
+		return title
+	case 2:
+		title = colorizeShowAndTell(title)
+		title = labelYCStartups(title)
+		return title
+	default:
+		return title
+	}
 }
 
 func formatDomain(domain string) string {
@@ -31,7 +44,7 @@ func paren(text string) string {
 	return " (" + text + ")"
 }
 
-func formatShowAndTell(title string) string {
+func labelShowAndTell(title string) string {
 	title = strings.ReplaceAll(title, "Ask HN:", reverse("Ask HN:"))
 	title = strings.ReplaceAll(title, "Show HN:", reverse("Show HN:"))
 	title = strings.ReplaceAll(title, "Tell HN:", reverse("Tell HN:"))
@@ -39,13 +52,21 @@ func formatShowAndTell(title string) string {
 	return title
 }
 
+func colorizeShowAndTell(title string) string {
+	title = strings.ReplaceAll(title, "Ask HN:", "[purple]"+"Ask HN:"+"[-]")
+	title = strings.ReplaceAll(title, "Show HN:", "[maroon]"+"Show HN:"+"[-]")
+	title = strings.ReplaceAll(title, "Tell HN:", "[navy]"+"Tell HN:"+"[-]")
+	title = strings.ReplaceAll(title, "Launch HN:", "[green]"+"Launch HN:"+"[-]")
+	return title
+}
+
 func reverse(text string) string {
 	return "[::r]" + text + "[-:-:-]"
 }
 
-func formatYCStartups(title string) string {
+func labelYCStartups(title string) string {
 	startYear := 05
-	endYear := 21
+	endYear := 22
 
 	for i := startYear; i <= endYear; i++ {
 		year := fmt.Sprintf("%02d", i)
@@ -53,15 +74,15 @@ func formatYCStartups(title string) string {
 		summer := "(YC S" + year + ")"
 		winter := "(YC W" + year + ")"
 
-		title = strings.ReplaceAll(title, summer, orange(summer))
-		title = strings.ReplaceAll(title, winter, orange(winter))
+		title = strings.ReplaceAll(title, summer, orange("YC S"+year))
+		title = strings.ReplaceAll(title, winter, orange("YC W"+year))
 	}
 
 	return title
 }
 
 func orange(text string) string {
-	return "[orange:black]" + text + "[-:-:-]"
+	return "[orange:" + black + "]" + text + "[-:-:-]"
 }
 
 func GetSecondaryText(points int, author string, time string, comments int) string {