+++ view/view.go
@@ -5,33 +5,37 @@ import (
 	"clx/constants/panels"
 	"clx/constants/submissions"
 	constructor "clx/constructors"
-	"clx/pages"
 	"clx/core"
+	"clx/pages"
 	text "github.com/MichaelMure/go-term-text"
 	"gitlab.com/tslocum/cview"
 	"strconv"
 	"time"
 )
 
+const (
+	black = "#0c0c0c"
+)
+
 func SetHackerNewsHeader(m *core.MainView, screenWidth int, category int) {
 	switch category {
 	case submissions.FrontPage:
-		base := "[black:orange:]   [Y[] [::b]Hacker News[::-]  new | ask | show"
+		base := "[" + black+":orange:]   [Y[] [::b]Hacker News[::-]  new | ask | show"
 		offset := -26
 		header := appendWhitespace(base, offset, screenWidth)
 		m.Header.SetText(header)
 	case submissions.New:
-		base := "[black:orange:]   [Y[] [::b]Hacker News[::-]  [white]new[black::] | ask | show"
+		base := "[" + black+":orange:]   [Y[] [::b]Hacker News[::-]  [white]new[" + black+"::] | ask | show"
 		offset := -42
 		header := appendWhitespace(base, offset, screenWidth)
 		m.Header.SetText(header)
 	case submissions.Ask:
-		base := "[black:orange:]   [Y[] [::b]Hacker News[::-]  new | [white]ask[black::] | show"
+		base := "[" + black+":orange:]   [Y[] [::b]Hacker News[::-]  new | [white]ask[" + black+"::] | show"
 		offset := -42
 		header := appendWhitespace(base, offset, screenWidth)
 		m.Header.SetText(header)
 	case submissions.Show:
-		base := "[black:orange:]   [Y[] [::b]Hacker News[::-]  new | ask | [white]show[black::]"
+		base := "[" + black+":orange:]   [Y[] [::b]Hacker News[::-]  new | ask | [white]show[" + black+"::]"
 		offset := -42
 		header := appendWhitespace(base, offset, screenWidth)
 		m.Header.SetText(header)
@@ -51,17 +55,17 @@ func appendWhitespace(base string, offset int, screenWidth int) string {
 func SetHelpScreenHeader(m *core.MainView, screenWidth int, category int) {
 	switch category {
 	case help.Info:
-		base := "[black:#82aaff:]   [^] [::b]circumflex[::-]   keymaps | settings"
+		base := "[" + black+":#82aaff:]   [^] [::b]circumflex[::-]   keymaps | settings"
 		offset := -26
 		header := appendWhitespace(base, offset, screenWidth)
 		m.Header.SetText(header)
 	case help.Keymaps:
-		base := "[black:#82aaff:]   [^] [::b]circumflex[::-]   [white]keymaps[black::] | settings"
+		base := "[" + black+":#82aaff:]   [^] [::b]circumflex[::-]   [white]keymaps[" + black+"::] | settings"
 		offset := -42
 		header := appendWhitespace(base, offset, screenWidth)
 		m.Header.SetText(header)
 	case help.Settings:
-		base := "[black:#82aaff:]   [^] [::b]circumflex[::-]   keymaps | [white]settings[black::]"
+		base := "[" + black+":#82aaff:]   [^] [::b]circumflex[::-]   keymaps | [white]settings[" + black+"::]"
 		offset := -42
 		header := appendWhitespace(base, offset, screenWidth)
 		m.Header.SetText(header)